refactor(blocks-construct): code refactor in file manager feature and resolve the sonarqube issue
