+++ src/features/file-manager/components/common-grid-view/common-grid-view.tsx
@@ -1,6 +1,7 @@
 import React, { useState, useCallback, useMemo } from 'react';
 import { Button } from '@/components/ui/button';
 import { SkeletonGrid } from '../file-manager-grid-view-skeleton/file-manager-grid-view-skeleton';
+import { GridViewCard } from '../grid-view-card/grid-view-card';
 
 interface BaseFile {
   id: string;
@@ -42,101 +43,6 @@ interface BaseGridViewProps<T extends BaseFile> {
   processFiles?: (files: T[]) => T[];
   currentFolderId?: string;
 }
-
-interface BaseCardProps<T extends BaseFile> {
-  file: T;
-  onFilePreview?: (file: T) => void;
-  onNavigateToFolder?: (folderId: string) => void;
-  renderActions: (file: T) => React.ReactNode;
-  IconComponent: React.ComponentType<{ className?: string }>;
-  iconColor: string;
-  backgroundColor: string;
-}
-
-const CommonCard = <T extends BaseFile>({
-  file,
-  onFilePreview,
-  onNavigateToFolder,
-  renderActions,
-  IconComponent,
-  iconColor,
-  backgroundColor,
-}: BaseCardProps<T>) => {
-  const isFolder = file.fileType === 'Folder';
-
-  const handleCardClick = (e: React.MouseEvent) => {
-    const target = e.target as HTMLElement;
-    const isActionButton = target.closest('[data-action-button="true"]');
-
-    if (!isActionButton) {
-      e.preventDefault();
-
-      if (isFolder && onNavigateToFolder) {
-        onNavigateToFolder(file.id);
-      } else if (!isFolder && onFilePreview) {
-        onFilePreview(file);
-      }
-    }
-  };
-
-  const containerClasses =
-    'group relative bg-background rounded-lg border border-border hover:bg-muted/50 hover:border-neutral-50 hover:shadow-md transition-all duration-200 cursor-pointer';
-  const contentClasses = isFolder
-    ? 'p-3 flex items-center space-x-3'
-    : 'p-6 flex flex-col items-center text-center space-y-4';
-
-  const iconContainerClasses = `${isFolder ? 'w-8 h-8' : 'w-20 h-20'} flex items-center justify-center rounded-lg ${isFolder ? backgroundColor : ''}`;
-  const iconClasses = `${isFolder ? 'w-5 h-5' : 'w-10 h-10'} ${iconColor} rounded-lg`;
-
-  const renderFolderLayout = () => (
-    <div className="flex items-center justify-between">
-      <div className="flex-1 min-w-0">
-        <h3 className="text-sm font-medium text-high-emphasis truncate" title={file.name}>
-          {file.name}
-        </h3>
-      </div>
-      <section data-action-button="true" className="flex-shrink-0" aria-label="File actions">
-        {renderActions(file)}
-      </section>
-    </div>
-  );
-
-  const renderFileLayout = () => (
-    <div className="flex items-center justify-between space-x-2 mt-2">
-      <div className="flex items-center space-x-2 flex-1 min-w-0">
-        <div className={`w-8 h-8 rounded-lg flex items-center justify-center ${backgroundColor}`}>
-          <IconComponent className={`w-4 h-4 ${iconColor}`} />
-        </div>
-        <div className="flex-1 min-w-0">
-          <h3 className="text-sm font-medium text-high-emphasis truncate" title={file.name}>
-            {file.name}
-          </h3>
-        </div>
-      </div>
-      <section data-action-button="true" className="flex-shrink-0" aria-label="File actions">
-        {renderActions(file)}
-      </section>
-    </div>
-  );
-
-  return (
-    <button
-      className={containerClasses}
-      onClick={handleCardClick}
-      aria-label={isFolder ? `Open folder ${file.name}` : `Preview ${file.name}`}
-      type="button"
-    >
-      <div className={contentClasses}>
-        <div className={iconContainerClasses}>
-          <IconComponent className={iconClasses} />
-        </div>
-        <div className={isFolder ? 'flex-1' : 'w-full'}>
-          {isFolder ? renderFolderLayout() : renderFileLayout()}
-        </div>
-      </div>
-    </button>
-  );
-};
 
 export const CommonGridView = <T extends BaseFile>({
   onFilePreview,
@@ -264,7 +170,7 @@ export const CommonGridView = <T extends BaseFile>({
                     const { iconColor, backgroundColor } = getFileTypeInfo(file.fileType);
 
                     return (
-                      <CommonCard
+                      <GridViewCard
                         key={`folder-${file.id}-${file.name}`}
                         file={file}
                         onNavigateToFolder={onNavigateToFolder}
@@ -291,7 +197,7 @@ export const CommonGridView = <T extends BaseFile>({
                     const { iconColor, backgroundColor } = getFileTypeInfo(file.fileType);
 
                     return (
-                      <CommonCard
+                      <GridViewCard
                         key={`file-${file.id}-${file.name}`}
                         file={file}
                         onNavigateToFolder={onNavigateToFolder}
