+++ src/features/file-manager/utils/mock-data.ts
@@ -1,4 +1,4 @@
-import { FileType } from '../types/file-manager.type';
+import { FileType, FilterOption } from '../types/file-manager.type';
 import { IFileDataWithSharing, IFileTrashData, SharedUser } from './file-manager';
 
 interface BaseFileDefinition {
@@ -660,3 +660,26 @@ export const folderContents: Record<string, IFileTrashData[]> = {
     },
   ],
 };
+
+export const getFileTypeFilters = (t: (key: string) => string): FilterOption[] => [
+  {
+    value: 'Folder',
+    label: t('FOLDER'),
+  },
+  {
+    value: 'File',
+    label: t('FILE'),
+  },
+  {
+    value: 'Image',
+    label: t('IMAGE'),
+  },
+  {
+    value: 'Audio',
+    label: t('AUDIO'),
+  },
+  {
+    value: 'Video',
+    label: t('VIDEO'),
+  },
+];
